# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Это API для управления сделками по продуктам.",
        "title": "Stawberry API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/audit/logs": {
            "get": {
                "description": "Retrieve audit trail entries with time range filtering and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audit"
                ],
                "summary": "Get audit logs",
                "parameters": [
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "Start time in RFC3339 format (default: 24h ago)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "End time in RFC3339 format (default: now)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by user ID",
                        "name": "uid",
                        "in": "query"
                    },
                    {
                        "maximum": 500,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Items per page (default 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns paginated audit logs",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
...
        "/guest/offers": {
            "post": {
                "description": "Allows sending an offer for a product on behalf of a guest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "guest"
                ],
                "summary": "Send a guest offer",
                "parameters": [
                    {
                        "description": "Guest offer data",
                        "name": "offer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/guestoffer.GuestPostOfferReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Offer accepted and forwarded",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid guest offer data",
                        "schema": {
                            "$ref": "#/definitions/apperror.Error"
                        }
                    },
                    "404": {
                        "description": "Store or product not found",
                        "schema": {
                            "$ref": "#/definitions/apperror.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apperror.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperror.Error": {
            "type": "object",
            "properties": {
                "errCode": {
                    "type": "string"
                },
                "errMsg": {
                    "type": "string"
                },
                "wrappedErr": {}
            }
        },
...
        "guestoffer.GuestPostOfferReq": {
            "type": "object",
            "required": [
                "currency",
                "guest_email",
                "guest_name",
                "guest_phone",
                "offer_price",
                "product_id",
                "store_id"
            ],
            "properties": {
                "currency": {
                    "type": "string"
                },
                "guest_email": {
                    "type": "string"
                },
                "guest_name": {
                    "type": "string"
                },
                "guest_phone": {
                    "type": "string"
                },
                "offer_price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                },
                "store_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token for authentication. Format: \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}